server/requirements.txt
asyncpg==0.28.0
alembic==1.12.0
bcrypt==4.0.1
fastapi~=0.100.0
greenlet==3.0.0rc2
gunicorn==21.2.0
SQLAlchemy~=2.0.20
pydantic~=2.3.0
python-dotenv~=1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytest==7.4.0
pytest-asyncio==0.21.0
pytz==2023.3
passlib~=1.7.0
uvicorn==0.23.0
websockets==11.0